refactor: rename Setlists to SetlistSearch
