+++ src/App.css
@@ -229,11 +229,6 @@ button:hover {
   list-style-type: none;
 }
 
-#instructions {
-  text-align: left;
-  margin-left: 5%;
-}
-
 img {
   width: 80%;
 }
