+++ src/tests/containers/SetlistSearch/Setlists.test.js
@@ -1,13 +1,13 @@
 import { mount } from 'enzyme';
 import React, { useState as useStateMock } from 'react';
-import Setlists from '../../../containers/SetlistSearch/Setlists';
+import SetlistSearch from '../../../containers/SetlistSearch/SetlistSearch';
 
 jest.mock('react', () => ({
   ...jest.requireActual('react'),
   useState: jest.fn(),
 }));
 
-describe('Setlists Component', () => {
+describe('SetlistSearch Component', () => {
   const setState = jest.fn();
 
   beforeEach(() => {
@@ -19,14 +19,14 @@ describe('Setlists Component', () => {
   });
 
   xit('renders an input field for the setlistId with a default value', () => {
-    const wrapper = mount(<Setlists />);
+    const wrapper = mount(<SetlistSearch />);
     const input = wrapper.find('#setlistId');
 
     expect(input.props().value).toEqual('');
   });
 
   xit('updates setlistId state with value in input field', () => {
-    const wrapper = mount(<Setlists />);
+    const wrapper = mount(<SetlistSearch />);
     const input = wrapper.find('#setlistId');
 
     input.simulate('change', { target: { value: 'abc1234' } });
